// lib/interfaces.ts
export interface Marker<T> {
  offset: number;
  openWithOffset?: (offset: number) => T;
  getParentOffset?: () => number;
}